% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_crs.R
\name{expect_crs}
\alias{expect_crs}
\alias{expect_crs}
\title{Expect coordinate reference systems}
\usage{
expect_crs(object, epsg = NULL, proj4string = NULL)
}
\arguments{
\item{object}{object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}}

\item{epsg}{A character for EPSG codes}

\item{proj4string}{A character string of projection arguments}
}
\description{
Expect coordinate reference systems
}
\examples{
sfc <- sf::st_sfc(sf::st_point(c(0, 0)),
sf::st_point(c(1, 1)))
sf <- sf::st_sf(a = 1:2, geom = sfc)
sf::st_crs(sf) <- 4326
expect_crs(sf, 4326)
expect_crs(sf, proj4string = "+proj=longlat +datum=WGS84 +no_defs")
}
