% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_units.R
\name{expect_units}
\alias{expect_units}
\title{Expect a units}
\usage{
expect_units(object, units = NULL)
}
\arguments{
\item{object}{object of class units}

\item{units}{expect units name}
}
\description{
Expect a units
}
\examples{
library(units)
x <- as_units(1, "m")
expect_units(x, "m")

expect_units(set_units(x, "km"), "km")
}
