% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_geometry.R
\name{expect_geometry}
\alias{expect_geometry}
\alias{expect_geom_type}
\title{Expect geometry type}
\usage{
expect_geom_type(object, type = c("POINT", "LINESTRING", "POLYGON",
  "MULTIPOINT", "MULTISTRING", "MULTIPOLYGON", "GEOMETRYCOLLECTION"))
}
\arguments{
\item{object}{object of class \href{sf::sf}{sf} or \code{\link[sf]{sfc}}}

\item{type}{test equality geometry type (character)}
}
\description{
Expect geometry type
}
\examples{
library(sf)
x <- st_point(c(0, 0))
expect_geom_type(x, "POINT")
}
